+++ lib/circuit_switch/notification.rb
@@ -10,7 +10,7 @@ module CircuitSwitch
   class CalledNotification < CircuitSwitchNotification
     def to_message(called_path:)
       if ::CircuitSwitch.config.with_backtrace
-        "#{message}\ncalled_path: #{called_path}\n#{StacktraceModifier.call(backtrace: backtrace)}"
+        "#{message}\ncalled_path: #{called_path}\n#{backtrace.join("\n")}"
       else
         message
       end