+++ test/lib/workers/reporter_test.rb
@@ -8,7 +8,7 @@ class ReporterTest < Test::Unit::TestCase
   end
 
   def test_reporter_calls_config_reporter
-    message = CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, run: false)
+    message = CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, stacktrace: [], run: false)
     assert_equal(
       CircuitSwitch::CircuitSwitch.last.message,
       message
@@ -20,7 +20,7 @@ class ReporterTest < Test::Unit::TestCase
       0,
       CircuitSwitch::CircuitSwitch.all.size
     )
-    CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, run: false)
+    CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, stacktrace: [], run: false)
     assert_equal(
       true,
       CircuitSwitch::CircuitSwitch.exists?(report_count: 1, key: called_path)
@@ -32,7 +32,7 @@ class ReporterTest < Test::Unit::TestCase
       0,
       CircuitSwitch::CircuitSwitch.all.size
     )
-    CircuitSwitch::Reporter.new.perform(key: 'test', limit_count: 10, called_path: called_path, run: false)
+    CircuitSwitch::Reporter.new.perform(key: 'test', limit_count: 10, called_path: called_path, stacktrace: [], run: false)
     assert_equal(
       true,
       CircuitSwitch::CircuitSwitch.exists?(report_count: 1, key: 'test')
@@ -40,12 +40,12 @@ class ReporterTest < Test::Unit::TestCase
   end
 
   def test_reporter_updates_circuit_switch_report_count_when_switch_exists
-    CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, run: false)
+    CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, stacktrace: [], run: false)
     assert_equal(
       1,
       CircuitSwitch::CircuitSwitch.find_by(key: called_path, caller: called_path).report_count
     )
-    CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, run: false)
+    CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, stacktrace: [], run: false)
     assert_equal(
       2,
       CircuitSwitch::CircuitSwitch.find_by(key: called_path, caller: called_path).report_count
@@ -53,12 +53,12 @@ class ReporterTest < Test::Unit::TestCase
   end
 
   def test_reporter_updates_circuit_switch_report_count_with_key_when_switch_exists
-    CircuitSwitch::Reporter.new.perform(key: 'test', limit_count: 10, called_path: called_path, run: false)
+    CircuitSwitch::Reporter.new.perform(key: 'test', limit_count: 10, called_path: called_path, stacktrace: [], run: false)
     assert_equal(
       1,
       CircuitSwitch::CircuitSwitch.find_by(key: 'test').report_count
     )
-    CircuitSwitch::Reporter.new.perform(key: 'test', limit_count: 10, called_path: called_path, run: false)
+    CircuitSwitch::Reporter.new.perform(key: 'test', limit_count: 10, called_path: called_path, stacktrace: [], run: false)
     assert_equal(
       2,
       CircuitSwitch::CircuitSwitch.find_by(key: 'test', caller: called_path).report_count
@@ -67,7 +67,7 @@ class ReporterTest < Test::Unit::TestCase
 
   def test_reporter_raises_active_record_record_not_found_when_reported_and_no_switch_exists
     assert_raise ActiveRecord::RecordNotFound do
-      CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, run: true)
+      CircuitSwitch::Reporter.new.perform(key: nil, limit_count: 10, called_path: called_path, stacktrace: [], run: true)
     end
   end
 end